+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerView.swift
@@ -25,7 +25,7 @@ public struct MessageComposerView<Factory: ViewFactory>: View, KeyboardReadable 
         viewFactory: Factory,
         viewModel: MessageComposerViewModel? = nil,
         channelController: ChatChannelController,
-        messageController: ChatMessageController?,
+        messageController: ChatMessageController? = nil,
         quotedMessage: Binding<ChatMessage?>,
         editedMessage: Binding<ChatMessage?>,
         onMessageSent: @escaping () -> Void
@@ -197,6 +197,7 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
     @Injected(\.colors) private var colors
     @Injected(\.fonts) private var fonts
     @Injected(\.images) private var images
+    @Injected(\.utils) private var utils
 
     var factory: Factory
     @Binding var text: String
@@ -254,6 +255,10 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
         }
 
         return textHeight
+    }
+    
+    var inputPaddingsConfig: PaddingsConfig {
+        utils.composerConfig.inputPaddingsConfig
     }
 
     public var body: some View {
@@ -340,8 +345,9 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
             }
             .frame(height: textFieldHeight)
         }
-        .padding(.vertical, shouldAddVerticalPadding ? 8 : 0)
-        .padding(.leading, 8)
+        .padding(.vertical, shouldAddVerticalPadding ? inputPaddingsConfig.vertical : 0)
+        .padding(.leading, inputPaddingsConfig.leading)
+        .padding(.trailing, inputPaddingsConfig.trailing)
         .background(composerInputBackground)
         .overlay(
             RoundedRectangle(cornerRadius: TextSizeConstants.cornerRadius)