+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageAvatarView.swift
@@ -2,7 +2,6 @@
 // Copyright © 2023 Stream.io Inc. All rights reserved.
 //
 
-import NukeUI
 import StreamChat
 import SwiftUI