+++ Sources/StreamChatSwiftUI/ChatChannelList/ChatChannelListItem.swift
@@ -219,7 +219,7 @@ extension ChatChannel {
 
     public var lastMessageText: String? {
         if let latestMessage = latestMessages.first {
-            return "\(latestMessage.author.name ?? latestMessage.author.id): \(latestMessage.textContent ?? latestMessage.adjustedText)"
+            return "\(latestMessage.author.name ?? latestMessage.author.id): \(textContent(for: latestMessage))"
         } else {
             return nil
         }
@@ -229,6 +229,13 @@ extension ChatChannel {
         !currentlyTypingUsersFiltered(
             currentUserId: InjectedValues[\.chatClient].currentUserId
         ).isEmpty && config.typingEventsEnabled
+    }
+    
+    public var shouldShowOnlineIndicator: Bool {
+        !lastActiveMembers.filter { member in
+            member.isOnline && member.id != InjectedValues[\.chatClient].currentUserId
+        }
+        .isEmpty
     }
 
     public var subtitleText: String {
@@ -248,6 +255,47 @@ extension ChatChannel {
             return InjectedValues[\.utils].dateFormatter.string(from: lastMessageAt)
         } else {
             return ""
+        }
+    }
+    
+    private func textContent(for previewMessage: ChatMessage) -> String {
+        if let attachmentPreviewText = attachmentPreviewText(for: previewMessage) {
+            return attachmentPreviewText
+        }
+        if let textContent = previewMessage.textContent, !textContent.isEmpty {
+            return textContent
+        }
+        return previewMessage.adjustedText
+    }
+    
+    /// The message preview text in case it contains attachments.
+    /// - Parameter previewMessage: The preview message of the channel.
+    /// - Returns: A string representing the message preview text.
+    private func attachmentPreviewText(for previewMessage: ChatMessage) -> String? {
+        guard let attachment = previewMessage.allAttachments.first else {
+            return nil
+        }
+        let text = previewMessage.textContent ?? previewMessage.text
+        switch attachment.type {
+        case .audio:
+            let defaultAudioText = L10n.Channel.Item.audio
+            return "🎧 \(text.isEmpty ? defaultAudioText : text)"
+        case .file:
+            guard let fileAttachment = previewMessage.fileAttachments.first else {
+                return nil
+            }
+            let title = fileAttachment.payload.title
+            return "📄 \(title ?? text)"
+        case .image:
+            let defaultPhotoText = L10n.Channel.Item.photo
+            return "📷 \(text.isEmpty ? defaultPhotoText : text)"
+        case .video:
+            let defaultVideoText = L10n.Channel.Item.video
+            return "📹 \(text.isEmpty ? defaultVideoText : text)"
+        case .giphy:
+            return "/giphy"
+        default:
+            return nil
         }
     }
 }