+++ Sources/StreamChatSwiftUI/ChatChannelList/ChatChannelListView.swift
@@ -21,7 +21,7 @@ public struct ChatChannelListView<Factory: ViewFactory>: View {
     private var handleTabBarVisibility: Bool
 
     public init(
-        viewFactory: Factory,
+        viewFactory: Factory = DefaultViewFactory.shared,
         viewModel: ChatChannelListViewModel? = nil,
         channelListController: ChatChannelListController? = nil,
         title: String = "Stream Chat",
@@ -176,7 +176,7 @@ public struct ChatChannelListContentView<Factory: ViewFactory>: View {
 
     private var viewFactory: Factory
     @ObservedObject private var viewModel: ChatChannelListViewModel
-    @StateObject private var channelHeaderLoader = ChannelHeaderLoader()
+    @ObservedObject private var channelHeaderLoader = InjectedValues[\.utils].channelHeaderLoader
     private var onItemTap: (ChatChannel) -> Void
 
     public init(