+++ CHANGELOG.md
@@ -3,7 +3,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### ⚠️ Important
+
+- Dependencies are no longer exposed (this includes Nuke and SwiftyGif). If you were using those dependencies we were exposing, you would need to import them manually. This is due to our newest addition supporting Module Stable XCFramework, see more below in the "Added" section. If you encounter any SPM-related problems, be sure to reset the package caches.
+
+### ✅ Added
+- Add message preview with attachments in channel list
+- Add support for pre-built XCFramework
+- Config for composer text input paddings
+- Config for left alignment of messages
+
 ### 🔄 Changed
+- Made some `ChannelList` and `MessageListView` parameters optional
 
 # [4.39.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.39.0)
 _October 06, 2023_